Add VMSA to measurement calculation
Given that VMSAs are separated out into their own consideration from the
launch-update-data-map, they must be accounted for as an extra step in
recomputing the measurement from the entire launch configuration map.

This specifies default values for VMSAs to make reference values more
compact when possible, but does not require that VMSAs begin with these
values.