Add a header separator
A number of tools choke without a clear end to the YAML header.

See https://github.com/cabo/kramdown-rfc/issues/259 for
a discussion of why this is `...` (and not `---`).
The important summary there is that this causes VSCode to be happy,
but `---` would be better for GitHub (this does nothing for GitHub.
Using `---` causes kramdown-rfc to emit a confusing warning,
so until that is fixed, we'll stick with `...`.